Fix: renamed entity saver to local storage
