+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:realm_roller/assets_handlers/entities_saver/entities_saver.dart';
+import 'package:realm_roller/assets_handlers/local_storage/local_storage.dart';
 import 'package:realm_roller/assets_handlers/route_observer.dart';
 import 'package:realm_roller/pages/general/main_page/main_page.dart';
 import 'package:realm_roller/theme/theme_data.dart';
