+++ lib/pages/general/menu_pages/saved/saved_entities_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:realm_roller/assets_handlers/entities_saver/entities_saver.dart';
+import 'package:realm_roller/assets_handlers/local_storage/local_storage.dart';
 import 'package:realm_roller/assets_handlers/route_observer.dart';
 import 'package:realm_roller/custom_widgets/cards/card_factory.dart';
